Support pre-signed URLs
This feauture is built on top of:
https://github.com/openshift/assisted-service/blob/master/docs/enhancements/image-service-cloud-authentication.md
https://github.com/openshift/assisted-service/blob/master/docs/enhancements/short-url-enhancement.md

The flow to get URL:

GET http://{service}/api/v1/sources/{source-id}/image-url

This will generate the URL which could be used for 4hours by user to download the OVA,
then the token will be invalidated.

Each source keep private key to sign&verify the URL token.

Signed-off-by: Ondra Machacek <omachace@redhat.com>